+++ src/test/java/org/hosh/runtime/CompilerTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willReturn;
 
-import java.util.Collections;
 import java.util.List;
 
 import org.hosh.runtime.Compiler.CompileError;
@@ -13,7 +12,6 @@ import org.hosh.runtime.Compiler.Program;
 import org.hosh.runtime.Compiler.Statement;
 import org.hosh.spi.Command;
 import org.hosh.spi.CommandWrapper;
-import org.hosh.spi.State;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -23,8 +21,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.StrictStubs.class)
 public class CompilerTest {
-	@Mock
-	private State state;
 	@Mock(stubOnly = true)
 	private Command command;
 	@Mock(stubOnly = true)
@@ -51,31 +47,21 @@ public class CompilerTest {
 	@Test
 	public void commandWithVariableExpansionWithSpace() {
 		given(commandResolver.tryResolve("cd")).willReturn(command);
-		given(state.getVariables()).willReturn(Collections.singletonMap("DIR", "/tmp"));
 		Program program = sut.compile("cd ${DIR}");
 		assertThat(program.getStatements()).hasSize(1);
 		List<Statement> statements = program.getStatements();
 		assertThat(statements.get(0).getCommand()).isSameAs(command);
-		assertThat(statements.get(0).getArguments()).contains("/tmp");
+		assertThat(statements.get(0).getArguments()).contains("${DIR}");
 	}
 
 	@Test
 	public void commandWithVariableExpansionNoSpace() {
 		given(commandResolver.tryResolve("echo")).willReturn(command);
-		given(state.getVariables()).willReturn(Collections.singletonMap("DIR", "/tmp"));
 		Program program = sut.compile("echo ${DIR}/aaa");
 		assertThat(program.getStatements()).hasSize(1);
 		List<Statement> statements = program.getStatements();
 		assertThat(statements.get(0).getCommand()).isSameAs(command);
-		assertThat(statements.get(0).getArguments()).contains("/tmp", "/aaa");
-	}
-
-	@Test
-	public void commandWithUnknownVariableExpansion() {
-		given(commandResolver.tryResolve("cd")).willReturn(command);
-		assertThatThrownBy(() -> sut.compile("cd ${DIR}"))
-				.isInstanceOf(CompileError.class)
-				.hasMessage("line 1: unknown variable DIR");
+		assertThat(statements.get(0).getArguments()).contains("${DIR}", "/aaa");
 	}
 
 	@Test