+++ src/test/java/org/hosh/runtime/InterpreterTest.java
@@ -1,9 +1,11 @@
 package org.hosh.runtime;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -22,23 +24,23 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.StrictStubs.class)
 public class InterpreterTest {
 	private Map<String, String> variables = new HashMap<>();
-	@Mock
+	private List<String> args = new ArrayList<>();
+	@Mock(stubOnly = true)
 	private State state;
 	@Mock(stubOnly = true)
 	private Terminal terminal;
 	@Mock(name = "out", stubOnly = true)
 	private Channel out;
-	@Mock
-	private Program program;
-	@Mock
-	private Statement statement;
 	@Mock(stubOnly = true)
-	private List<String> args;
+	private Program program;
+	@Mock(stubOnly = true)
+	private Statement statement;
 	@Mock
 	private Command command;
 	@Mock(extraInterfaces = StateAware.class)
@@ -96,5 +98,42 @@ public class InterpreterTest {
 		then(terminalAwareCommand).should().run(args, out, out);
 		then((TerminalAware) terminalAwareCommand).should().setTerminal(terminal);
 		then(terminalAwareCommand).shouldHaveNoMoreInteractions();
+	}
+
+	@Test
+	public void plainArguments() throws Exception {
+		args.add("file");
+		given(state.getVariables()).willReturn(variables);
+		given(command.run(Mockito.any(), Mockito.any(), Mockito.any())).willReturn(ExitStatus.success());
+		given(program.getStatements()).willReturn(Arrays.asList(statement));
+		given(statement.getCommand()).willReturn(command);
+		given(statement.getArguments()).willReturn(args);
+		sut.eval(program);
+		then(command).should().run(Arrays.asList("file"), out, out);
+		then(command).shouldHaveNoMoreInteractions();
+	}
+
+	@Test
+	public void resolvePresentVariables() throws Exception {
+		args.add("${VARIABLE}");
+		variables.put("VARIABLE", "1");
+		given(state.getVariables()).willReturn(variables);
+		given(command.run(Mockito.any(), Mockito.any(), Mockito.any())).willReturn(ExitStatus.success());
+		given(program.getStatements()).willReturn(Arrays.asList(statement));
+		given(statement.getCommand()).willReturn(command);
+		given(statement.getArguments()).willReturn(args);
+		sut.eval(program);
+		then(command).should().run(Arrays.asList("1"), out, out);
+		then(command).shouldHaveNoMoreInteractions();
+	}
+
+	@Test
+	public void refuseAbsentVariables() throws Exception {
+		args.add("${VARIABLE}");
+		given(state.getVariables()).willReturn(variables);
+		given(program.getStatements()).willReturn(Arrays.asList(statement));
+		given(statement.getCommand()).willReturn(command);
+		given(statement.getArguments()).willReturn(args);
+		assertThatThrownBy(() -> sut.eval(program)).isInstanceOf(IllegalStateException.class).hasMessageContaining("unknown variable: VARIABLE");
 	}
 }