+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.0.18] - 2018-10-20
 - requires Java 11
+- lazy resolving variables at runtime
 - removing glob pattern from 'find'
   (the idea is to use pipelines to filter paths)
 - 'ps' command